Upadted script ../ush/trpsfprv.sh

+++ util/ush/trpsfprv.sh
@@ -22,45 +22,51 @@ set -x
 ###################################
 # Copy Grib files
 ###################################
-cp ${COMIN}/gfs.$cycle.pgrb2.1p00.f024  .
-export err=$?
-if [[ $err -ne 0 ]] ; then
-   echo " File gfs.$cycle.pgrb2.1p00.f024 does not exist."
-   exit $err
-fi
 
-cp ${COMIN}/gfs.$cycle.pgrb2.1p00.f048  .
-export err=$?
-if [[ $err -ne 0 ]] ; then
-   echo " File gfs.$cycle.pgrb2.1p00.f048 does not exist."
-   exit $err
-fi
+# Check for PGRB (GRIB1) files exist or not
 
-$CNVGRIB -g21 gfs.$cycle.pgrb2.1p00.f024 pgrbf24
-export err=$?
-if [[ $err -ne 0 ]] ; then
-   echo " CNVGRIB failed to convert GRIB2 to GRIB1 "
-   exit $err
-fi
-
-$GRBINDEX pgrbf24 pgrbif24
-err1=$?
-
-$CNVGRIB -g21 gfs.$cycle.pgrb2.1p00.f048 pgrbf48
-export err=$?
-if [[ $err -ne 0 ]] ; then
-   echo " CNVGRIB failed to convert GRIB2 to GRIB1 "
-   exit $err
-fi
-$GRBINDEX pgrbf48 pgrbif48
-err2=$?
-
-tot=`expr $err1 + $err2`
-if test "$tot" -ne 0
+if test -f pgrbf24 
 then
-  msg="File not yet available in com"
-  postmsg "$jlogfile" "$msg"
-  err_exit
+   echo " File pgrbf24 GRIB1 exist."
+   break
+else
+   cp ${COMIN}/gfs.$cycle.pgrb2.1p00.f024  .
+   export err=$?
+   if [[ $err -ne 0 ]] ; then
+       echo " File gfs.$cycle.pgrb2.1p00.f024 does not exist."
+       exit $err
+   fi
+
+   cp ${COMIN}/gfs.$cycle.pgrb2.1p00.f048  .
+   export err=$?
+   if [[ $err -ne 0 ]] ; then
+       echo " File gfs.$cycle.pgrb2.1p00.f048 does not exist."
+      exit $err
+   fi
+   $CNVGRIB -g21 gfs.$cycle.pgrb2.1p00.f024 pgrbf24
+   export err=$?
+   if [[ $err -ne 0 ]] ; then
+       echo " CNVGRIB failed to convert GRIB2 to GRIB1 "
+       exit $err
+   fi
+   $GRBINDEX pgrbf24 pgrbif24
+   err1=$?
+
+   $CNVGRIB -g21 gfs.$cycle.pgrb2.1p00.f048 pgrbf48
+   export err=$?
+   if [[ $err -ne 0 ]] ; then
+      echo " CNVGRIB failed to convert GRIB2 to GRIB1 "
+      exit $err
+   fi
+   $GRBINDEX pgrbf48 pgrbif48
+   err2=$?
+   tot=`expr $err1 + $err2`
+   if test "$tot" -ne 0
+   then
+     msg="File not yet available in com"
+     postmsg "$jlogfile" "$msg"
+     err_exit
+   fi
 fi
 
 # cp $PARMshared/graph_trpsfprv_ft08.$cycle trpsfprv_ft08.$cycle
